Confirm hooks

+++ components/data-table.tsx
@@ -26,6 +26,7 @@ import {
 } from "@/components/ui/table"
 import React from "react"
 import { Trash } from "lucide-react";
+import { useConfirm } from "@/hooks/use-confirm";
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
@@ -41,6 +42,11 @@ export function DataTable<TData, TValue>({
   filterKey,
   onDelete,disabled
 }: DataTableProps<TData, TValue>) {
+
+  const [ConfirmDialog,confirm]=useConfirm(
+    "Are you sure",
+    "You are about to perform a bulk delete!"
+  )
 
     const [sorting, setSorting] = React.useState<SortingState>([]);
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
@@ -70,6 +76,7 @@ export function DataTable<TData, TValue>({
 
   return (
    <div>
+    <ConfirmDialog />
     <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -86,9 +93,13 @@ export function DataTable<TData, TValue>({
             size="sm"
             variant="outline"
             className="ml-auto text-xs font-normal" 
-            onClick={()=>{
-                onDelete(table.getFilteredSelectedRowModel().rows);
+            onClick={async()=>{
+              const ok=await confirm();
+
+                if(ok){
+                  onDelete(table.getFilteredSelectedRowModel().rows);
                 table.resetRowSelection();
+                }
             }}
            
           >
